requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.43.0
jinja2==3.1.5
python-multipart==0.0.20
aiofiles==24.1.0
pydantic==2.10.4
python-dotenv==1.0.1
